+++ CHANGELOG.md
@@ -1,8 +1,6 @@
 # Changelog
 
 ## [Unreleased]
-
-
 - #332 Можно отправить во фронт информативное сообщение об ошибке поиска в одну строку: throw new SearchRequestException(errorMessage, HttpStatus.BAD_REQUEST);
 - #282 Добавлен View для Aircraft(в т.ч. клиент feign). DTO и API перенесено в Common. DTO теперь использует Mapper. Методы createAircraft() и updateAircraft() в API Теперь возвращает DTO вместо Entity;
 - #327 Рефакторинг Security. Аутентифкация на стороне security модуля, авторизация на стороне gateway
@@ -24,6 +22,7 @@
 - #320 Фронт стартовой страницы с поиском
 - #321 Фронт для поиска свободных мест и отображения результатов, рефакторинг FlightDTO, FlightSeat, Search, SearchResult
 - #321 Рефакторинг DestinationDTO, FlightSeatDTO, фронт Destination, обработка FeignException во фронте
+- #192 Добавлены тесты на ендпоинт api/seats/aircraft/{aircraftId}
 - #286 Добавлен фронт для Flight
 
 ## [1.0.7] - 8.09.2023